requirements.txt
# The Grand FuSa Agent — 最小依赖（按需取消注释或增补）

# httpx>=0.27.0
# openai>=1.0.0
# pydantic>=2.0.0
# pydantic-settings>=2.0.0

# Document generation (YAML -> MD/DOCX/XLSX)
pyyaml
python-docx
openpyxl
